[issue-24] Moved methods and added context menu with dummy methods.
